Refactor invoice widgets to share reusable traits
